--- python/cudf/cudf/utils/__init__.py
@@ -1,1 +1,0 @@
-from cudf.utils.utils import hash_cudf_object
